reformat tests
