+++ x/programs/cmd/simulator/src/context.rs
@@ -1,16 +1,7 @@
 use crate::codec::base64_encode;
-use crate::step::SimulatorError;
-use crate::{param::Param, Id};
-use base64::{engine::general_purpose::STANDARD as b64, Engine};
-use borsh::BorshDeserialize;
-use serde::{Deserialize, Deserializer, Serialize, Serializer};
-use std::{
-    io::{BufRead, BufReader, Write},
-    path::Path,
-    process::{Child, Command, Stdio},
-};
-use thiserror::Error;
-use wasmlanche_sdk::{Address, ExternalCallError};
+use crate::Id;
+use serde::{Serialize, Serializer};
+use wasmlanche_sdk::Address;
 
 #[derive(Clone, Debug, PartialEq, Default)]
 #[non_exhaustive]
