+++ x/programs/rust/examples/counter/src/lib.rs
@@ -38,53 +38,41 @@ fn get_value_internal(context: &Context<StateKeys>, of: Address) -> Count {
 
 #[cfg(test)]
 mod tests {
-    use simulator::{Endpoint, Step, TestContext};
+    use simulator::{build_simulator, TestContext};
     use wasmlanche_sdk::Address;
 
     const PROGRAM_PATH: &str = env!("PROGRAM_PATH");
 
     #[test]
     fn init_program() {
-        let mut simulator = simulator::ClientBuilder::new().try_build().unwrap();
+        let mut simulator = build_simulator().unwrap();
 
-        simulator
-            .run_step(&Step::create_program(PROGRAM_PATH))
-            .unwrap();
+        simulator.create_program(PROGRAM_PATH).unwrap();
     }
 
     #[test]
     fn increment() {
-        let mut simulator = simulator::ClientBuilder::new().try_build().unwrap();
+        let mut simulator = build_simulator().unwrap();
 
         let bob = Address::new([1; 33]);
 
-        let counter_id = simulator
-            .run_step(&Step::create_program(PROGRAM_PATH))
-            .unwrap()
-            .id;
+        let counter_id = simulator.create_program(PROGRAM_PATH).unwrap().id;
 
         let test_context = TestContext::from(counter_id);
 
-        simulator
-            .run_step(&Step {
-                endpoint: Endpoint::Execute,
-                method: "inc".into(),
-                max_units: 1000000,
-                params: vec![test_context.clone().into(), bob.into(), 10u64.into()],
-            })
-            .unwrap();
+        simulator.execute(
+            "inc".into(),
+            vec![test_context.clone().into(), bob.into(), 10u64.into()],
+            1000000,
+        );
 
         let value = simulator
-            .run_step(&Step {
-                endpoint: Endpoint::ReadOnly,
-                method: "get_value".into(),
-                max_units: 0,
-                params: vec![test_context.into(), bob.into()],
-            })
+            .read("get_value".into(), vec![test_context.into(), bob.into()])
             .unwrap()
             .result
             .response::<u64>()
             .unwrap();
+
         assert_eq!(value, 10);
     }
 }
